benchmarks settings refactor
