+++ benchmarks/settings.hpp
@@ -58,6 +58,26 @@ namespace rfaas::benchmark {
     static Settings deserialize(std::istream & in);
   };
 
+  struct Options {
+
+    std::string json_config;
+    std::string device_database;
+    std::string executors_database;
+    std::string output_stats;
+    bool verbose;
+    std::string fname;
+    std::string flib;
+    int input_size;
+
+    int cores;
+
+    bool rdma_type;
+    int pause;
+    int read_size;
+
+  };
+
+  Options options(int argc, char ** argv);
 }
 
 #endif
