+++ cmake/benchmarks.cmake
@@ -2,13 +2,14 @@
 add_library(benchmarks STATIC benchmarks/settings.cpp)
 target_include_directories(benchmarks PRIVATE $<TARGET_PROPERTY:rfaaslib,INTERFACE_INCLUDE_DIRECTORIES>)
 target_include_directories(benchmarks PRIVATE $<TARGET_PROPERTY:spdlog::spdlog,INTERFACE_INCLUDE_DIRECTORIES>)
+target_include_directories(benchmarks SYSTEM PRIVATE $<TARGET_PROPERTY:cxxopts::cxxopts,INTERFACE_INCLUDE_DIRECTORIES>)
 target_link_libraries(benchmarks INTERFACE rfaaslib)
 
-add_executable(warm_benchmarker benchmarks/warm_benchmark.cpp benchmarks/warm_benchmark_opts.cpp)
-add_executable(parallel_invocations benchmarks/parallel_invocations.cpp benchmarks/parallel_invocations_opts.cpp)
-add_executable(cold_benchmarker benchmarks/cold_benchmark.cpp benchmarks/cold_benchmark_opts.cpp)
-add_executable(cpp_interface benchmarks/cpp_interface.cpp benchmarks/cpp_interface_opts.cpp)
-add_executable(rma benchmarks/rma.cpp benchmarks/cpp_interface_opts.cpp)
+add_executable(warm_benchmarker benchmarks/warm_benchmark.cpp)
+add_executable(parallel_invocations benchmarks/parallel_invocations.cpp)
+add_executable(cold_benchmarker benchmarks/cold_benchmark.cpp)
+add_executable(cpp_interface benchmarks/cpp_interface.cpp)
+add_executable(rma benchmarks/rma.cpp)
 set(tests_targets "rma" "warm_benchmarker" "cold_benchmarker" "parallel_invocations" "cpp_interface")
 foreach(target ${tests_targets})
   add_dependencies(${target} cxxopts::cxxopts)
